Fix: session 생성 and hook 이름
